+++ internal/solr/admin.go
@@ -4,7 +4,9 @@ import (
 	"context"
 	"fmt"
 	"github.com/PuerkitoBio/goquery"
+	"github.com/rs/zerolog/log"
 	"io"
+	"io/ioutil"
 	"net/http"
 	"net/url"
 	"strings"
@@ -23,7 +25,6 @@ func NewConnectionAndAdminClient(uri string, client *http.Client) (*AdminClient,
 	if err != nil {
 		return nil, err
 	}
-	uri = uri + "/solr/"
 
 	conn := &Connection{Uri: uri, httpClient: client}
 	bp := conn.formatBasePath()
@@ -101,4 +102,48 @@ func AdminPing(uri string) (string, string, error) {
 	}
 
 	return specversion, solrhome, nil
+}
+
+func UpdateSolrSchema(uri string, corename string, schemafilename string) error {
+	bytes, err := ioutil.ReadFile(schemafilename)
+	if err != nil {
+		log.Err(err)
+		return err
+	}
+
+	//TODO: basepathを使う
+	uri = uri + "/solr/" + corename + "/schema"
+	req, err := http.NewRequest(http.MethodPost, uri, strings.NewReader(string(bytes)))
+	if err != nil {
+		log.Err(err)
+		return err
+	}
+	req.Header.Set("Content-Type", "Content-type:application/json")
+
+	fmt.Println("Post schema file")
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		log.Err(err)
+		return err
+	}
+	// deferでクローズ処理
+	defer func(Body io.ReadCloser) {
+		err := Body.Close()
+		if err != nil {
+
+		}
+	}(resp.Body)
+	// Bodyの内容を読み込む
+	body, _ := io.ReadAll(resp.Body)
+
+	if resp.StatusCode != 200 {
+		log.Error().Msgf("error: response code is %d", resp.StatusCode)
+		fmt.Println("Error:")
+		fmt.Print(string(body))
+		return fmt.Errorf("error: response code is %d", resp.StatusCode)
+	}
+
+	fmt.Println("success")
+	return nil
 }
